+++ package.json
@@ -13,10 +13,11 @@
     "style:fix": "pnpm style --fix",
     "format": "prettier . --write",
     "preview": "vite preview",
-    "predeploy": "pnpm run build",
+    "NOpredeploy": "pnpm run build",
     "deploy": "gh-pages -d dist",
     "generate-pwa-assets": "pwa-assets-generator",
     "cdf-scrap": "tsx src/scripts/cdf-scrap.ts",
+    "vb-scrap": "tsx src/scripts/vb-scrap.ts",
     "pdf-parse": "tsx src/scripts/pdf-parse.ts"
   },
   "dependencies": {
@@ -39,7 +40,7 @@
   "devDependencies": {
     "@nabla/vite-plugin-eslint": "^2.0.4",
     "@supercharge/promise-pool": "^3.2.0",
-    "@tanstack/eslint-plugin-query": "^5.58.1",
+    "@tanstack/eslint-plugin-query": "^5.59.1",
     "@types/d3-graphviz": "^2.6.10",
     "@types/lodash": "^4.17.10",
     "@types/node": "^22.7.4",
@@ -48,7 +49,7 @@
     "@types/react-dom": "^18.3.0",
     "@typescript-eslint/eslint-plugin": "^7.18.0",
     "@typescript-eslint/parser": "^7.18.0",
-    "@vite-pwa/assets-generator": "^0.2.4",
+    "@vite-pwa/assets-generator": "^0.2.6",
     "@vitejs/plugin-react-swc": "^3.7.1",
     "eslint": "^8.57.1",
     "eslint-config-prettier": "^9.1.0",
